split deleteFiles into UI and logic parts
