+++ internal/restic/parallel.go
@@ -3,7 +3,9 @@ package restic
 import (
 	"context"
 
+	"github.com/restic/restic/internal/backend"
 	"github.com/restic/restic/internal/debug"
+	"github.com/restic/restic/internal/ui/progress"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -48,5 +50,45 @@ func ParallelList(ctx context.Context, r Lister, t FileType, parallelism uint, f
 		wg.Go(worker)
 	}
 
+	return wg.Wait()
+}
+
+// ParallelRemove deletes the given fileList of fileType in parallel
+// if callback returns an error, then it will abort.
+func ParallelRemove(ctx context.Context, repo Repository, fileList IDSet, fileType FileType, report func(id ID, err error) error, bar *progress.Counter) error {
+	fileChan := make(chan ID)
+	wg, ctx := errgroup.WithContext(ctx)
+	wg.Go(func() error {
+		defer close(fileChan)
+		for id := range fileList {
+			select {
+			case fileChan <- id:
+			case <-ctx.Done():
+				return ctx.Err()
+			}
+		}
+		return nil
+	})
+
+	bar.SetMax(uint64(len(fileList)))
+
+	// deleting files is IO-bound
+	workerCount := repo.Connections()
+	for i := 0; i < int(workerCount); i++ {
+		wg.Go(func() error {
+			for id := range fileChan {
+				h := backend.Handle{Type: fileType, Name: id.String()}
+				err := repo.Backend().Remove(ctx, h)
+				if report != nil {
+					err = report(id, err)
+				}
+				if err != nil {
+					return err
+				}
+				bar.Add(1)
+			}
+			return nil
+		})
+	}
 	return wg.Wait()
 }
